safari-534.53-branch/LayoutTests/sputnik/Conformance/12_Statement/12.7_The_continue_Statement/S12.7_A5_T3-expected.txt

CONSOLE MESSAGE: line 81: SyntaxError: Parse error
S12.7_A5_T3

PASS Expected parsing failure

TEST COMPLETE
